enable for safari + improves
