+++ src/app/(volume)/layout.jsx
@@ -1,0 +1,7 @@
+export default function Layout({ children }) {
+  return (
+    <html className="bg-black">
+      <body>{children}</body>
+    </html>
+  );
+}
