+++ src/app/(volume)/volume/page.jsx
@@ -4,11 +4,7 @@ import { getVolumeTransactionsPerDay } from '@/features/volume/actions';
 export const revalidate = 60;
 
 export default async function Page() {
-  const [transactions] = await Promise.all([getVolumeTransactionsPerDay()]);
+  const transactions = await getVolumeTransactionsPerDay();
 
-  return (
-    <html className="bg-black">
-      <Volume transactions={transactions} />
-    </html>
-  );
+  return <Volume transactions={transactions} />;
 }
