+++ src/features/volume/components/VolumeChart.jsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Area,
   AreaChart,
@@ -13,32 +15,7 @@ import CustomTooltip from '@/components/Recharts/CustomTooltip';
 import { formatCurrency } from '@/lib/utils';
 
 const processData = (transactions) => {
-  const transactionsFrom = transactions.filter(
-    (transaction) => new Date(transaction.date) > new Date('2024-03-25'),
-  );
-  const groupedByWeek = transactionsFrom.reduce(
-    (acc, { date, totalReceivedAmount }) => {
-      const dateObj = new Date(date);
-      const dayOfWeek = dateObj.getDay(); // Get day of week (0 for Sunday, 6 for Saturday)
-      const startOfWeek = new Date(dateObj);
-      startOfWeek.setDate(
-        dateObj.getDate() - dayOfWeek + (dayOfWeek === 0 ? -6 : 1),
-      ); // Adjust to Monday as the first day of the week
-      startOfWeek.setHours(0, 0, 0, 0); // Normalize to start of the day
-
-      // Format the start of the week as "MM-DD-YYYY"
-      const weekKey = `${(startOfWeek.getMonth() + 1).toString().padStart(2, '0')}/${startOfWeek.getDate().toString().padStart(2, '0')}/${startOfWeek.getFullYear()}`;
-
-      if (!acc[weekKey]) {
-        acc[weekKey] = 0;
-      }
-      acc[weekKey] += totalReceivedAmount;
-      return acc;
-    },
-    {},
-  );
-
-  return Object.entries(groupedByWeek).map(([name, amt]) => ({
+  return Object.entries(transactions).map(([name, amt]) => ({
     name, // Week start date in "MM-DD-YYYY" format
     amt, // Sum of amounts for the week
   }));
